refactor: refactor adapter
