+++ packages/iceworks-server/src/interface/base.ts
@@ -78,10 +78,6 @@ export interface IProject {
 export interface IBaseModule {
   project: IProject;
   storage: any;
-  isAvailable: boolean;
-  title: string;
-  description?: string;
-  cover?: string;
 }
 
 /**
