+++ packages/iceworks-server/src/lib/adapter-ice-scripts-1-kit-2/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "adapter-ice-scripts-1-kit-2"
+}
