+++ packages/iceworks-server/src/lib/adapter-ice-scripts-2-kit-3/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "adapter-ice-scripts-2-kit-3"
+}
