+++ packages/iceworks-server/src/lib/adapter/locales/en-US.json
@@ -1,0 +1,1 @@
+{}
