+++ packages/iceworks-server/src/lib/adapter/locales/zh-CN.json
@@ -1,0 +1,1 @@
+{}
