+++ packages/iceworks-server/src/lib/adapter/modules/configuration/index.ts
@@ -1,21 +1,12 @@
 import * as path from 'path';
-import { getCLIConf, setCLIConf } from '../utils/cliConf';
-import { ICLIConf, IProject, IConfigurationModule, IConfParam } from '../../../interface';
+import { getCLIConf, setCLIConf } from '../../utils/cliConf';
+import { ICLIConf, IProject, IConfigurationModule, IConfParam } from '../../../../interface';
 import { CLI_CONF } from './const';
-import config from '../config';
-
-const { title,  description, cover, isAvailable } = config['configuration'];
 
 export default class Configuration implements IConfigurationModule {
-  public readonly title: string = title;
-  public readonly description: string = description;
-  public readonly cover: string = cover;
-  public readonly isAvailable: boolean = isAvailable;
   public project: IProject;
   public storage: any;
-
   public readonly cliConfPath: string;
-
   private cliConfFilename = 'ice.config.js';
 
   constructor(params: {project: IProject; storage: any; }) {
