+++ packages/iceworks-server/src/lib/adapter/modules/menu/index.ts
@@ -6,22 +6,15 @@ import generate from '@babel/generator';
 import * as t from '@babel/types';
 import * as uid from 'uid';
 
-import formatCodeFromAST from '../formatCodeFromAST';
-import { IMenu, IProject, IMenuModule, IMenuOptions } from '../../../interface';
-import config from '../config';
+import formatCodeFromAST from '../../utils/formatCodeFromAST';
+import { IMenu, IProject, IMenuModule, IMenuOptions } from '../../../../interface';
 
 const MENU_CONFIG_VARIABLE = 'asideMenuConfig';
-const { title,  description, cover, isAvailable } = config['menu'];
 
 export default class Menu implements IMenuModule {
-  public readonly title: string = title;
-  public readonly description: string = description;
-  public readonly cover: string = cover;
-  public readonly isAvailable: boolean = isAvailable;
   public readonly projectPath: string;
   public readonly project: IProject;
   public readonly storage: any;
-
   public readonly path: string;
 
   constructor(params: {project: IProject; storage: any; }) {
