+++ packages/iceworks-server/src/lib/adapter/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "adapter-ice-scripts-2-kit-2"
+}
