+++ packages/iceworks-server/src/lib/plugin/project-manager/app.ts
@@ -11,7 +11,7 @@ import * as orderBy from 'lodash.orderby';
 import * as arrayMove from 'array-move';
 import camelCase from 'camelCase';
 import storage from '../../storage';
-import * as adapter from '../../adapter';
+import adapter from '../../adapter';
 import { IProject, IMaterialScaffold, IPanel, IBaseModule } from '../../../interface';
 import getTarballURLByMaterielSource from '../../getTarballURLByMaterielSource';
 import downloadAndExtractPackage from '../../downloadAndExtractPackage';
@@ -70,14 +70,17 @@ class Project implements IProject {
   }
 
   private loadAdapter() {
-    for (const [name, Module] of Object.entries(adapter)) {
+    for (const [name, config] of Object.entries(adapter)) {
       const project: IProject = clone(this);
       delete project.adapter;
 
-      const adapterModule = new Module({ project, storage });
-      this.adapter[camelCase(name)] = adapterModule;
+      const Module = config.module;
+      if (Module) {
+        const adapterModule = new Module({ project, storage });
+        this.adapter[camelCase(name)] = adapterModule;
+      }
 
-      const {title, description, cover, isAvailable } = adapterModule;
+      const {title, description, cover, isAvailable } = config;
       this.panels.push({
         name,
         title,
